Retarget minSdkVersion to something slightly modern-ish.
Anything before Sdk 21 results in a multidex packaging error with a 64k
limit that I don't care to try and explain because I'd have to look it
all up again.

Long story short:  Anything below Sdk 21 is pretty much impossible to
build for in Flutter unless you don't pull in any other projects.  Now
that we're wiring up to Firebase that's a library and BOOM we're over
the limit.